src/main/java/com/example/java_tamuna_metonidzee/Flights.java
package com.example.java_tamuna_metonidzee;

import java.io.Serializable;
import java.util.Date;

/**
 * Domain model
 *
 * @author Julian Jupiter
 *
 */
public class Flights implements Serializable {

    private int id;
    private String direction;
    private Date dateTime;
    private int numberOfSeats;
    private int totalPrice;

    public int getId() {
        return id;
    }

    public void setId(int id) {
        this.id = id;
    }

    public String getDirection() {
        return direction;
    }

    public void setDirection(String direction) {
        this.direction = direction;
    }

    public Date getDateTime() {
        return dateTime;
    }

    public void setDateTime(Date dateTime) {
        this.dateTime = dateTime;
    }

    public int getNumberOfSeats() {
        return numberOfSeats;
    }

    public void setNumberOfSeats(int numberOfSeats) {
        this.numberOfSeats = numberOfSeats;
    }

    public int getTotalPrice() {
        return totalPrice;
    }

    public void setTotalPrice(int totalPrice) {
        this.totalPrice = totalPrice;
    }

}
